UI finished, graphics are next
UI working as a charm. Tested all the possible combinations I can think
of and killed all the bugs in the code. Not as pretty as it should be,
but it's solid while processing.
Next step is the creation of .svg graphics for the truss model and
results.

Also Toby the kitty came to life. :)