Started on .svg graphics
Started using rasem gem to create .svg graphics.
The file Test_rasem.rb is a standalone version for model graphics output
including all the necessary new methods.  The next step is to divide
this file into different files that will output graphics for the model,
model with displacements and model with internal forces.