Alpha release (Toby approved)
Finished with Ruby Truss.

The program was tested on multiple examples and should work for any kind
of trusses. There are three simple school examples (Example 1, 2, 3) and
one complex Crane truss example.

There is still a lot of room for improvement, but that will have to wait
for another time.